requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
torchreid>=0.2.5
moviepy>=1.0.3
numpy>=1.24.0
gdown>=4.7.0
tensorboard>=2.14.0
pyyaml>=6.0
flask>=3.0.0
